Linter passes.
