+++ package.json
@@ -57,9 +57,11 @@
     "test": "mocha test/*.js"
   },
   "dependencies": {
-    "google-proto-files": "^0.13.1",
-    "google-gax": "^0.14.2",
     "extend": "^3.0.0",
+    "google-gax": "^0.14.2",
+    "google-proto-files": "^0.13.1",
+    "lodash.merge": "^4.6.0",
+    "protobufjs": "^6.8.0",
     "pumpify": "^1.3.5",
     "stream-events": "^1.0.1",
     "through2": "^2.0.3"
