+++ samples/package.json
@@ -24,6 +24,7 @@
     "@google-cloud/nodejs-repo-tools": "1.4.17",
     "ava": "0.21.0",
     "proxyquire": "1.8.0",
-    "sinon": "3.2.0"
+    "sinon": "3.2.0",
+    "uuid": "^3.1.0"
   }
 }
